backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
Pillow==10.3.0
django-filter==24.2
bcrypt==4.1.2
django-simple-history==3.7.0
